# specs/v1/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "nanocld",
    "description": "",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": ""
    },
    "version": "0.1.3"
  },
  "paths": {
...
    "/git_repositories": {
      "get": {
        "tags": [
          "git_repository"
        ],
        "summary": "List all git repository",
        "description": "List all git repository\n",
        "operationId": "list_git_repository",
        "responses": {
          "200": {
            "description": "Array of git_repository",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GitRepositoryItem"
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "git_repository"
        ],
        "summary": "Create new git repository",
        "description": "Create new git repository\n",
        "operationId": "create_git_repository",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitRepositoryPartial"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Fresh created git_repository",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitRepositoryItem"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error"
          },
          "404": {
            "description": "Namespace name not valid"
          },
          "422": {
            "description": "The provided payload is not valid"
          }
        },
        "deprecated": false
      }
    },
    "/git_repositories/{name}": {
      "delete": {
        "tags": [
          "git_repository"
        ],
        "summary": "Delete git repository by it's name",
        "description": "Delete git repository by it's name\n",
        "operationId": "delete_git_repository_by_name",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Name of git repository",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Number of entry deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgDeleteGeneric"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error"
          },
          "404": {
            "description": "Namespace name not valid"
          }
        },
        "deprecated": false
      }
    },
    "/git_repositories/{name}/build": {
      "post": {
        "tags": [
          "git_repository"
        ],
        "summary": "Transform a git repository into an image",
        "description": "Transform a git repository into an image\n",
        "operationId": "build_git_repository_by_name",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Name of git repository",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "description": "Branch to build default to main branch",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Number of entry deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgDeleteGeneric"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error"
          },
          "404": {
            "description": "Namespace name not valid"
          }
        },
        "deprecated": false
      }
    },
    "/namespaces": {
      "get": {
        "tags": [
          "namespace"
        ],
        "summary": "List all namespace",
        "description": "List all namespace\n",
        "operationId": "list_namespace",
        "responses": {
          "200": {
            "description": "Array of namespace",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NamespaceItem"
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "namespace"
        ],
        "summary": "Create new namespace",
        "description": "Create new namespace\n",
        "operationId": "create_namespace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespacePartial"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "fresh created namespace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceItem"
                }
              }
            }
          },
          "400": {
            "description": "generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "the provided payload is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/namespaces/{name}": {
      "delete": {
        "tags": [
          "namespace"
        ],
        "summary": "Delete namespace by it's name",
        "description": "Delete namespace by it's name\n",
        "operationId": "delete_namespace_by_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "name of the namespace",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "database generic delete response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgDeleteGeneric"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/namespaces/{name}/inspect": {
      "get": {
        "tags": [
          "namespace"
        ],
        "summary": "Inspect namespace by it's name",
        "description": "Inspect namespace by it's name\n",
        "operationId": "inspect_namespace_by_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "name of the namespace",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Namespace found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceItem"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Api Error Structure that server send to client",
        "required": [
          "msg"
        ],
        "properties": {
          "msg": {
            "type": "string"
          }
        }
      },
...
      "GitRepositoryItem": {
        "type": "object",
        "description": "Git repository are used to have project definition to deploy cargo",
        "required": [
          "name",
          "url",
          "default_branch",
          "source"
        ],
        "properties": {
          "default_branch": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/GitRepositorySourceType"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "GitRepositoryPartial": {
        "type": "object",
        "description": "Partial Git repository",
        "required": [
          "url",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "GitRepositorySourceType": {
        "type": "string",
        "description": "Git repository source types",
        "enum": [
          "github",
          "gitlab",
          "local"
        ]
      },
      "NamespaceItem": {
        "type": "object",
        "description": "Namespace to encapsulate clusters",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "NamespacePartial": {
        "type": "object",
        "description": "Partial namespace",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}